codegen(axioms): promote approx= subtree to Real when any float literal is present

+++ verifiers/osmotic_pressure/rust-verifier/src/axioms.rs
@@ -18,13 +18,16 @@ use z3::{
 };
 #[cfg(feature = "smt")]
 use std::str::FromStr as _;
+#[cfg(feature = "smt")]
+#[allow(unused_imports)]
+use std::ops::{Add as _, Mul as _, Sub as _};
 
 #[cfg(feature = "smt")]
 pub fn assert_axioms(solver: &Solver) {
     // constraint C001-vant-hoff (approx-equality, tolerance 0.03)
     {
-        let lhs = Int::new_const("osmotic-pressure-pa_s");
-        let rhs = Int::new_const("vant-hoff-i_s").mul(&Int::new_const("molarity_s")).mul(&Real::from_rational(8314000, 1000000)).mul(&Int::new_const("temperature-k_s"));
+        let lhs = Real::new_const("osmotic-pressure-pa_s");
+        let rhs = Real::new_const("vant-hoff-i_s").mul(&Real::new_const("molarity_s")).mul(&Real::from_rational(8314000, 1000000)).mul(&Real::new_const("temperature-k_s"));
         let diff = lhs.sub(&rhs);
         let eps  = Real::from_rational(30000, 1000000);
         let neg_eps = Real::from_rational(-30000, 1000000);
